Fixes for issues id#5466, id#5467, id#5468. Instructions.cs: some adding of missing instruction prefix names, minor refactoring.
MDMethodElems.cs: method signature generated on demand to avoid read-order issues; In MethodSpec the signature of the MethParent must be hoisted up to this; ClassSpec objects are added to the TypeSpec table conditionally, to avoid duplicate entries.

MDTypeElems.cs: typeSpecAdded field is moved from TypeSpec to Type, so as to apply to ClassSpec objects.  Otherwise a TypeSpec wrapper class for ClassSpec would have been needed; In the same vein, the GenericParTypeSpec class that wraps a GenericParam for the TypeSpec list has to be referenced by the owning GenericParam so as no avoid duplicates losing object identity.

MetaDataOut.cs: conditional version of AddToTable for TypeSpec.

PEReader.cs: Extra clause in GetDataConstant to persist array initialization in .data